requirements.txt
agno
python-dotenv
google-generativeai
pandas 